+++ internal/mocks/doc.go
@@ -9,6 +9,7 @@ package mocks // import "github.com/slok/meterm/internal/mocks"
 // Render mocks.
 //go:generate mockery -output ./view/render -outpkg render -dir ../view/render -name Renderer
 //go:generate mockery -output ./view/render -outpkg render -dir ../view/render -name GaugeWidget
+//go:generate mockery -output ./view/render -outpkg render -dir ../view/render -name SinglestatWidget
 
 // Services mocks.
 //go:generate mockery -output ./service/metric -outpkg metric -dir ../service/metric -name Gatherer