+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Debug flag that will set a verbose logger (will break UI rendering but prints errors and infos).
 - Termdash render engine implementation for widgets.
+- Singlestat widget.
 - Gauge widget.
 - Main term application.
 - Fake metrics gatherer.